refactor(time tracker fetcher): rename the file name to match the class name
